Swapped out the timer GET for a signalr connection.

+++ server/RdtClient.Service/Services/Torrents.cs
@@ -169,7 +169,10 @@ namespace RdtClient.Service.Services
                     
                     var torrentPath = Path.Combine(settingDownloadFolder, torrent.RdName);
 
-                    Directory.Delete(torrentPath, true);
+                    if (Directory.Exists(torrentPath))
+                    {
+                        Directory.Delete(torrentPath, true);
+                    }
                 }
                 
                 if (deleteData)
